[Go] Sieve
After looking at the community solutions, it became clear that there are much better performing ways to do this.

I specifically like the use of a slice of booleans to 'mark' each number, without having to actually create a slice of the numbers themselves. You can just use the index.